+++ apps/ai-dial-admin/src/components/PublicationView/FileProperties/FilesList.tsx
@@ -7,7 +7,7 @@ import { ActionType } from '@/src/models/dial/publications';
 import Grid from '@/src/components/Grid/Grid';
 import { ApplicationRoute } from '@/src/types/routes';
 import { FILE_DOWNLOAD, FILE_PREVIEW, PREVIEW_EXTENSIONS } from '@/src/constants/file';
-import { getGridFileColumns, getGridFileData } from '@/src/components/FilesList/files-list';
+import { getGridFileColumns, getGridFileData } from '@/src/utils/files/grid-data';
 import {
   getDownloadOperation,
   getOpenInNewTabOperation,
