+++ apps/ai-dial-admin/src/constants/i18n.ts
@@ -497,6 +497,7 @@ export enum PublicationsI18nKey {
   ActionUnpublish = 'Publications.ActionUnpublish',
   Action = 'Publications.Action',
   Author = 'Publications.Author',
+  Runner = 'Publications.Runner',
   CreatedAt = 'Publications.CreatedAt',
   FolderStorage = 'Publications.FolderStorage',
   FilesListTitle = 'Publications.Files.FilesListTitle',
