+++ apps/ai-dial-admin/src/utils/files/get-extension.ts
@@ -1,0 +1,12 @@
+export const getNameExtensionFromFile = (input: string): { name: string; extension: string } => {
+  const lastUnderscoreIndex = input.lastIndexOf('.');
+
+  if (lastUnderscoreIndex === -1) {
+    return { name: input, extension: '' };
+  }
+
+  const name = input.substring(0, lastUnderscoreIndex);
+  const extension = input.substring(lastUnderscoreIndex);
+
+  return { name, extension };
+};
