+++ apps/ai-dial-admin/src/utils/is-asset-view.ts
@@ -1,0 +1,7 @@
+import { ApplicationRoute } from '@/src/types/routes';
+
+export const isAssetView = (view: ApplicationRoute): boolean => {
+  return (
+    view === ApplicationRoute.Prompts || view === ApplicationRoute.Files || view === ApplicationRoute.AssetsApplications
+  );
+};
